commit  3
boitte vitesse
fin crud

+++ src/main/java/com/spring/repository/CategorieRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.spring.models.Categorie;
 
-public interface CategorieRepository extends JpaRepository<Categorie, Long> {
+public interface CategorieRepository extends JpaRepository<Categorie, Integer> {
 
 }
